move controllers into pkg (#542)
